Use `int` instead of `long` for dices
There is no need to store a `long` number here as we are drawing
values between 1 and 6.

+++ api/src/main/java/com/monomalpoly/api/dice/Dice.java
@@ -3,16 +3,16 @@ package com.monomalpoly.api.dice;
 public class Dice {
 
     private final boolean isDouble;
-    private final long value;
+    private final int value;
     private final String message;
 
-    public Dice(long value, boolean isDouble, String message) {
+    public Dice(int value, boolean isDouble, String message) {
         this.value = value;
         this.isDouble = isDouble;
         this.message = message;
     }
 
-    public long getValue() {
+    public int getValue() {
         return value;
     }
